+++ libstreetmap/src/libcurl.cpp
@@ -12,7 +12,7 @@
 using namespace std;
 using boost::property_tree::ptree;
 using boost::property_tree::read_json;
-std::vector<double> weatherData;
+std::vector<int> weatherData;
 std::string currentWeather;
 
 typedef struct MyCustomStruct {
@@ -116,7 +116,6 @@ std::string chooseCity(std::string fileName){
 
 void loadCityWeatherData(std::string cityFile) {
     std::string cityName = chooseCity(cityFile);
-    //cout << cityName << endl;
     
     CURLcode res = curl_global_init(CURL_GLOBAL_ALL);
     if (res != CURLE_OK) {
